requirements.txt
altair==5.2.0
altgraph==0.17.3
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.1.0
click==8.1.3
colour==0.1.5
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
Flask==2.3.2
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.4
ipykernel==6.29.3
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json-normalize==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kaggle==1.6.6
kiwisolver==1.4.5
lxml==4.9.2
macholib==1.16.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.1
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.9.0
python-slugify==8.0.4
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
soupsieve==2.4.1
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.3.0
tkmacosx==1.0.5
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.1
torchtext==0.17.2
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.6
